trivial. Enhance js content replacer #3. Modify code structure
