+++ package.json
@@ -59,6 +59,7 @@
         "eslint-plugin-storybook": "^0.6.15",
         "eslint-plugin-testing-library": "^5.10.2",
         "fs-extra": "^11.1.1",
+        "glob": "^10.3.10",
         "jest": "^29.6.3",
         "jest-environment-jsdom": "^29.6.3",
         "react": "^18.2.0",
